requirements.txt
Django==4.1.5
djangorestframework==3.14.0
psycopg2-binary==2.9.5
python-decouple==3.6
drf-spectacular==0.25.1
django-cors-headers==3.13.0
